Rendergraph

+++ InsightGraphics/src/Graphics/GraphicsManager.cpp
@@ -31,7 +31,7 @@ namespace Insight
 		{
 			PixelFormatExtensions::Init();
 
-			m_sharedData.GraphicsAPI = GraphicsAPI::Vulkan;
+			m_sharedData.GraphicsAPI = GraphicsAPI::DX12;
 			currentGraphicsAPI = (int)m_sharedData.GraphicsAPI;
 
 			m_renderContext = RenderContext::New();
@@ -44,6 +44,10 @@ namespace Insight
 			{
 				return false;
 			}
+
+#ifdef RENDER_GRAPH_ENABLED
+			m_renderGraph.SetRenderContext(m_renderContext);
+#endif //#ifdef RENDER_GRAPH_ENABLED
 
 			m_renderpass.Create();
 
@@ -84,28 +88,35 @@ namespace Insight
 			const int previousGrapicsAPI = currentGraphicsAPI;
 			const char* graphicsAPIs[] = { "Vulkan", "DX12" };
 			IMGUI_VALID(ImGui::Begin(currentAPI.c_str()));
-			IMGUI_VALID(if (ImGui::ListBox("Graphcis API", &currentGraphicsAPI, graphicsAPIs, _countof(graphicsAPIs)))
+			if (ImGui::GetCurrentContext())
 			{
-				if (currentGraphicsAPI != previousGrapicsAPI)
+				if (ImGui::ListBox("Graphcis API", &currentGraphicsAPI, graphicsAPIs, _countof(graphicsAPIs)))
 				{
-					// New API
-					m_renderContext->GpuWaitForIdle();
-					m_renderpass.Destroy();
+					if (currentGraphicsAPI != previousGrapicsAPI)
+					{
+						// New API
+						m_renderContext->GpuWaitForIdle();
+						m_renderpass.Destroy();
 
-					m_renderContext->Destroy();
-					DeleteTracked(m_renderContext);
-					m_renderContext = nullptr;
-			
-					Window::Instance().Rebuild();
-			
-					m_sharedData.GraphicsAPI = (GraphicsAPI)currentGraphicsAPI;
-					m_renderContext = RenderContext::New();
-					m_renderContext->Init();
-					m_renderpass.Create();
-					Renderer::s_FrameCommandList.Reset();
-					return;
+						m_renderContext->Destroy();
+						DeleteTracked(m_renderContext);
+						m_renderContext = nullptr;
+
+						Window::Instance().Rebuild();
+
+						m_sharedData.GraphicsAPI = (GraphicsAPI)currentGraphicsAPI;
+						m_renderContext = RenderContext::New();
+						m_renderContext->Init();
+						m_renderpass.Create();
+						Renderer::s_FrameCommandList.Reset();
+
+#ifdef RENDER_GRAPH_ENABLED
+						m_renderGraph.SetRenderContext(m_renderContext);
+#endif //#ifdef RENDER_GRAPH_ENABLED
+						return;
+					}
 				}
-			})
+			}
 			IMGUI_VALID(ImGui::End());
 
 			m_renderpass.Render();
